Add uml diagrams

+++ src/main/java/org/example/fibonacci/Fibonacci.java
@@ -1,5 +1,5 @@
 package org.example.fibonacci;
 
 public interface Fibonacci {
-    int fibonacci(int num);
+    long fibonacci(int num);
 }
